requirements.txt
livekit-agents>=0.12.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-openai>=0.10.0
livekit-plugins-silero>=0.6.0
python-dotenv>=1.0.0
